+++ PCL.Core/IO/Net/Http/HostConnectionHandler.cs
