+++ PCL.Core/IO/Net/Http/HttpBasicExtension.cs
