+++ PCL.Core/IO/Net/Http/HttpContentExtension.cs
