+++ PCL.Core/IO/Net/Http/HttpCookieExtension.cs
