+++ PCL.Core/IO/Net/Http/HttpHeaderExtension.cs
