+++ PCL.Core/IO/Net/Http/HttpProxyManager.cs
