+++ PCL.Core/IO/Net/Http/HttpRequest.cs
