+++ PCL.Core/IO/Net/Http/HttpResponseExtension.cs
