+++ PCL.Core/IO/Net/Http/HttpRoute.cs
@@ -1,7 +1,7 @@
-﻿using System;
+using System;
 using System.Net.Http;
 
-namespace PCL.Core.IO.Net.Http.Server;
+namespace PCL.Core.IO.Net.Http;
 
 public class HttpRoute(HttpMethod method, string path) : Attribute
 {
